% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{agrégat_annuel}
\alias{agrégat_annuel}
\title{Affichage du tableau des agrégats des primes A et B, pour chaque année de période}
\usage{
agrégat_annuel(résultat, verbeux)
}
\arguments{
\item{résultat}{Résultat retourné par la fonction \link{test_prime}}

\item{verbeux}{[FALSE] Le résultat n'est affiché que si  \code{verbeux} vaut  \code{TRUE}}
}
\description{
Affichage du tableau des agrégats des primes A et B, pour chaque année de période
}
\examples{
agrégat_annuel(test_prime(prime_IAT, prime_B = prime_IFTS, Paie_I, verbeux = FALSE), verbeux = TRUE)
}
