% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ajuster.médiane.temps.complet}
\alias{ajuster.médiane.temps.complet}
\title{Ajuster la médiane des templs complets en heures}
\usage{
ajuster.m攼㸹diane.temps.complet()
}
\description{
Ajuster la médiane des templs complets en heures
}
