% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair.R
\name{altair}
\alias{altair}
\title{Altaïr: un paquet pour analyser les bases de paye XML.}
\description{
Altaïr: un paquet pour analyser les bases de paye XML.
}
\section{Démographie}{

effectifs : tableau des effectifs, ETP et EQTP

pyramide des âges : de 15 à 68 ans, verticale, hommes/femmes.
}

