% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{année_comparaison}
\alias{année_comparaison}
\title{Annee de comparaison avec les données nationales.}
\usage{
ann攼㸹e_comparaison(versant)
}
\arguments{
\item{versant}{Chaîne de caractères parmi "FPT", "FPH", "TIT_FPT", "TIT_FPH", "NONTIT_FPT", "NONTIT_FPH".}
}
\value{
Une liste dont les composantes sont :
       \describe{
        \item{\code{année}}{l'année de comparaison (entier).}
        \item{\code{pyr}}{la base de données nationales correspondante (data.table)}.
        }
}
\description{
Calcule l'année à laquelle la pyramide des âges va être comparée aux données nationales, pour un versant donné de la focntion publique.
}
