% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{calcul.nb.jours.mois}
\alias{calcul.nb.jours.mois}
\title{Calcul du nombre de jours dans le mois}
\usage{
calcul.nb.jours.mois(Mois, ann攼㸹e)
}
\description{
Calcul du nombre de jours dans le mois
}
