% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astreintes.R
\name{calcul_astreintes}
\alias{calcul_astreintes}
\title{Constrôle des astreintes pour les emplois de responsabilité supérieure}
\usage{
calcul_astreintes()
}
\description{
Vérification du non-paiement des astreintes à des emplois de directeur général.
}
\details{
\itemize{
\item{Filtre la paye en retenant les astreintes.}
\item{Retient les cumuls avec les NBI non nulles pour la catégorie A et un emploi ou un grade de directeur général} 
\item{Affiche le nombre d'agents détectés dans ces cas} 
\item{Crée la table \code{Controle_astreintes} ayant pour colonnes : \code{"Code.astreinte", "Libelle.astreinte", "Montant.astreinte", "Code.NBI", "Libelle.NBI", "Montant.NBI"}} 
\item{Crée la table \code{Cum_astreintes} ayant pour colonne : \code{Montant annuel astreintes}, et son total pour la période.}
\item{Sauve les bases \code{Controle_astreintes} et \code{libelles.astreintes} dans le répertoire \emph{Reglementation}}
\item{Affiche si besoin la table \code{Cum_astreintes} au format \code{\link{Tableau.vertical2}}}
}
}
\note{
Les emplois sont filtrés à l'aide de l'expression rationnelle Perl suivante, sans prise en compte de la casse : \cr
\code{d(?:\\.|ir)\\w*\\s*\\bg(?:\\.|\\w*n\\.?\\w*)\\s*\\b(?:des?)\\s*\\bs\\w.*}
}
\author{
Fabrice Nicol
}
