% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{calcul_rmpp}
\alias{calcul_rmpp}
\title{RMPP
Calcul de la rémunération moyenne des personnes en place}
\usage{
calcul_rmpp()
}
\description{
RMPP
Calcul de la rémunération moyenne des personnes en place
}
\note{
Filtrage : on enlève les personnels présents depuis moins d'un seuil de troncature (ex. 120 jours) dans l'année et les élus
(paramètre seuil.troncature) 
Filtrage pour l'étude des variations : on enlève les valeurs manquantes des variations, les centiles extrêmaux,
les rémunérations nettes négatives ou proche de zéro. On exige un statut explicite en fin de période.
Paramétrable par :
minimum.positif, quantile.cut
}
