% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilité.R
\name{calculer_indice_complexité}
\alias{calculer_indice_complexité}
\title{Diagnostic des tables de jointure}
\usage{
calculer_indice_complexit攼㸹()
}
\value{
NULL si les variables \code{Code, Libelle, Type} suffisent, sinon le tableau des multiplicités associées aux combinaisons de ces variables auxquelles il est impossible d'associer un compte unique.
}
\description{
Calcule le nombre minimum de variables nécessaires pour apparier les bases de paye et un tableau de correspondance entre codes de paye et comptabilité
}
