% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{charges.eqtp}
\alias{charges.eqtp}
\title{Tableau des coûts moyens de personnel par grade.}
\usage{
charges.eqtp(Base = Paie, grade = NULL, classe = NULL, service = NULL,
  libellés = NULL, agr = FALSE, période = NULL)
}
\arguments{
\item{Base}{Base des Paies, comportant pour l'ensemble de la période. Par défaut, base Paie, comportant
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Année}             
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Type}
         \item \code{Grade}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{Montant}}{réel}}}}}

\item{grade}{Vecteur de grades particuliers. Tous les grades en l'absence de spécification.}

\item{classe}{Liste de vecteurs de grades, chaque vecteur représentant une classe aggrégée, ou bien vecteur de chaîne de caractères représentant des expressions rationnelles sur les grades. Tous les grades en l'absence de spécification. La casse est ignorée pour les expressions rationnelles.}

\item{libellés}{Vecteur de libellés des agrégations de grades par expression régulière. Doit avoir la même dimension que le vecteur de regexp.}

\item{agr}{Booléen (défaut FALSE). Si TRUE, l'expression régulière précédente conduit à agréger les grades décrits par le vecteur d'expressions régulières précédent : une ligne par composante du vecteur.}

\item{période}{Vecteur des années considérées.}
}
\value{
Un tableau des charges moyennes de personnel par EQTP (coût employeur) mis en forme avec les grades en ligne et autant de colonnes numériques que d'années de période, plus une colonne de libellés.
}
\description{
Elabore un tableau des charges de personnel (coût) par grade et par année.
}
\examples{
charges.eqtp()
}
