% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{chemin}
\alias{chemin}
\title{Chemin complet d'un fichier dans le dossier \code{chemin.dossier.donnees}}
\usage{
chemin(fichier)
}
\arguments{
\item{fichier}{Nom de fichier}
}
\description{
Chemin complet d'un fichier dans le dossier \code{chemin.dossier.donnees}
}
