% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grades.catégories.R
\name{correspondance_grade_catégorie}
\alias{correspondance_grade_catégorie}
\title{Correspondance grade-catégorie}
\usage{
correspondance_grade_catégorie()
}
\value{
La \code{data.table grade.categorie} résultant de la lecture du fichier \bold{grades.catégories.csv} sous le répertoire \bold{Données}
}
\description{
Etablit la correspondance entre le grade et la catégorie statutaire à partir d'un tableau importé
}
\note{
Requiert l'utilisation d'une table de jointure importée \bold{grades.catégories.csv} sous le répertoire \bold{Données}.
A défaut, tente une association approximative à partir d'expressions rationnelles appliquées aux grades.
}
