% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptabilité.R
\name{correspondance_paye_budget}
\alias{correspondance_paye_budget}
\title{Correspondance paye-budget}
\usage{
correspondance_paye_budget()
}
\value{
La \code{data.table code.libellé} résultant de la lecture du fichier \bold{paye_budget.csv} sous le répertoire \bold{Données}
}
\description{
Etablit la correspondance entre paye et comptabilité administrative (comptes 64 et 65)
}
\note{
Requiert l'utilisation d'une table de jointure importee \bold{paye_budget.csv} encodée en UTF-8 sous le répertoire \bold{Données}.
A défaut, tente une association approximative à partir d'expressions rationnelles appliquées aux libellés de paye.
}
