% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astreintes.R
\name{cumul_astreintes_IHTS}
\alias{cumul_astreintes_IHTS}
\title{Contrôle du cumul des astreintes et des IHTS}
\usage{
cumul_astreintes_IHTS()
}
\description{
Détection du paiement le même mois d'indemnités d'astreintes et d'IHTS
}
\author{
Fabrice Nicol
}
