% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{distribution_smpt}
\alias{distribution_smpt}
\title{distribution_smpt}
\usage{
distribution_smpt(Filtre)
}
\arguments{
\item{Filtre}{Filtre fonctionnel booléen décrivant une sélection sur les lignes (caractéristiques des personnels)}
}
\value{
Quartiles du SMPT
}
\description{
distribution_smpt
}
