% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{eqtp.grade}
\alias{eqtp.grade}
\title{Tableau des EQTP par grade.}
\usage{
eqtp.grade(Base = Bulletins.paie, grade = NULL, classe = NULL,
  service = NULL, libellés = NULL, agr = FALSE, période = NULL,
  variation = FALSE, statut = NULL, catégorie = NULL)
}
\arguments{
\item{Base}{Base des bulletins de paye, comportant pour l'ensemble de la période
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Annee}
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Grade}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{quotite}}{réel entre 0 et 1}}}}.}

\item{grade}{Grade particulier. Tous les grades en l'absence de spécification.}

\item{classe}{Liste de vecteurs de grades, chaque vecteur représentant une classe aggrégée, ou bien vecteur de chaîne de caractères représentant des expressions rationnelles sur les grades. Tous les grades en l'absence de spécification. La casse est ignorée pour les expressions rationnelles.}

\item{service}{Services. Vecteur de chaînes de caractères exactes. Tous les services en l'absence de spécification.}

\item{libellés}{Vecteur de libellés des agrégations de grades par expression régulière. Doit avoir la même dimension que le vecteur de regexp.}

\item{agr}{Booléen (défaut FALSE). Si TRUE, l'expression régulière précédente conduit à agréger les grades décrits par le vecteur d'expressions régulières précédent : une ligne par composante du vecteur.}

\item{période}{Vecteur des années considérées.}

\item{variation}{Booléen Insérer une colonne des variations (défaut FALSE).}

\item{statut}{Restreindre le tableau au vecteur des statuts en paramètres. Expressions exactes. Tous statuts par défaut.}

\item{catégorie}{Categorie statutaire (vecteur de lettres parmi 'A', 'B', 'C'). Par défaut A, B, C ou indéterminée.}
}
\value{
Un tableau des effectifs par grade mis en forme avec les grades en ligne et autant de colonnes numériques que d'années de période, plus une colonne de libellés.
}
\description{
Elabore un tableau des équivalents temps plein travaillés par grade et par année.
}
\examples{
eqtp.grade()
}
