% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{file2Latin}
\alias{file2Latin}
\title{Conversion d'un fichier en ISO-8859-15}
\usage{
file2Latin(nom, encodage.in = "UTF-8")
}
\arguments{
\item{nom}{Chemin du fichier à encoder}

\item{encodage.in}{(= encodage.entrée) Encodage du fichier de lecture}
}
\value{
Lancement dun appel système à iconv -t ISO-8859-15
}
\description{
Conversion d'un fichier de UTF-8 en ISO-8859-15
}
