% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{file2utf8}
\alias{file2utf8}
\title{Conversion d'un fichier de ISO-8859-15 en UTF-8}
\usage{
file2utf8(nom, encodage.in = "ISO-8859-15")
}
\arguments{
\item{nom}{Chemin du fichier à encoder}

\item{encodage.in}{(= encodage.entrée) Encodage du fichier de lecture}
}
\value{
Lancement dun appel système à iconv -t UTF-8
}
\description{
Conversion d'un fichier de ISO-8859-15 en UTF-8
}
\author{
Fabrice Nicol
}
