% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{filtre}
\alias{filtre}
\title{Sélection du filtre correspondant à une chaîne de caractères associée}
\usage{
filtre(x)
}
\arguments{
\item{x}{Chaîne de caractères associée figurant dans la colonne \code{type} de la table \code{codes}}
}
\value{
Cette fonction renvoie un ensemble de valeurs fixes ou une expression régulière, sauf si x n'est pas dans la colonne \code{type} de la table \code{codes} (renvoie NULL)
}
\description{
Sélection du filtre correspondant à une chaîne de caractères associée
}
