% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{FR}
\alias{FR}
\title{Ajoute le séparateur des milliers et sans décimale en français}
\usage{
FR(x)
}
\arguments{
\item{x}{Vecteur de valeurs numériques}
}
\value{
Vecteur de chaînes de caractères modifiées
}
\description{
Ajoute le séparateur des milliers et sans décimale en français
}
\examples{
FR(1235345.23) #1 235 345,2
}
\author{
Fabrice Nicol
}
