% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{\%+\%}
\alias{\%+\%}
\title{Concaténation de chaîne de caractères}
\usage{
x \%+\% y
}
\arguments{
\item{x}{Vecteur de caractères}

\item{y}{Vecteur de caractères}
}
\value{
Vecteur de caractères concaténé.
}
\description{
Equivalent de paste0(,)
}
\examples{
x <- "abc"
y <- "def"
cat(x \%+\% y) # "abcdef"
}
\author{
Fabrice Nicol
}
