% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{\%s\%}
\alias{\%s\%}
\title{Prise en compte du pluriel}
\usage{
mot \%s\% N
}
\arguments{
\item{N}{Entier : pluriel si N > 1.}

\item{Mot}{Mot à pluraliser éventuellement}
}
\value{
Concaténation du mot et de "s" si N > 1
}
\description{
Prise en compte du pluriel
}
\examples{
cat("patient" \%s\% 2)
}
\author{
Fabrice Nicol
}
