% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importer_grades_catégories}
\alias{importer_grades_catégories}
\title{Importer la base externe des correspondances grades-catégories}
\usage{
importer_grades_catégories()
}
\description{
Importer la base externe des correspondances grades-catégories
}
