% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{longueur.non.na}
\alias{longueur.non.na}
\title{Longueur d'un vecteur, ou nombre de lignes d'une table, lorsque l'on a retiré les éléments ou les lignes NA}
\usage{
longueur.non.na(v)
}
\arguments{
\item{v}{Vecteur ou data.frame/data.table}
}
\value{
Nombre d'éléments ou de lignes.
}
\description{
Longueur d'un vecteur, ou nombre de lignes d'une table, lorsque l'on a retiré les éléments ou les lignes NA
}
