% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{net.eqtp.emploi}
\alias{net.eqtp.emploi}
\title{Tableau des rémunérations nettes moyennes par emploi.}
\usage{
net.eqtp.emploi(
  Base = Paie,
  grade = NULL,
  emploi = NULL,
  classe = NULL,
  service = NULL,
  libellés = NULL,
  agr = FALSE,
  periode = NULL,
  variation = FALSE,
  statut = NULL,
  categorie = NULL,
  exclure.codes = NULL,
  exclure.sft = FALSE,
  quotite.nulle = FALSE
)
}
\arguments{
\item{Base}{Base des bulletins de paye, comportant pour l'ensemble de la periode
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Annee}
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Emploi}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{quotite}}{réel entre 0 et 1}}}}.}

\item{emploi}{Emploi particulier. Tous les emplois en l'absence de spécification.}

\item{classe}{Liste de vecteurs d'emplois, chaque vecteur représentant une classe aggrégée, ou bien vecteur de chaîne de caractères représentant des expressions rationnelles sur les emplois. Tous les emplois en l'absence de spécification. La casse est ignorée pour les expressions rationnelles.}

\item{service}{Services. Vecteur de chaînes de caractères exactes. Tous les services en l'absence de spécification.}

\item{libellés}{Vecteur de libellés des agrégations d'emplois par expression régulière. Doit avoir la même dimension que le vecteur de regexp.}

\item{agr}{Booléen (défaut FALSE). Si TRUE, l'expression régulière précédente conduit à agréger les emplois décrits par le vecteur d'expressions régulières précédent : une ligne par composante du vecteur.}

\item{periode}{Vecteur des annees considérées.}

\item{variation}{Booléen Insérer une colonne des variations (défaut FALSE).}

\item{statut}{Restreindre le tableau au vecteur des statuts en paramètres. Expressions exactes. Tous statuts par défaut.}

\item{categorie}{Categorie statutaire (vecteur de lettres parmi 'A', 'B', 'C'). Par défaut A, B, C ou indéterminée.}

\item{exclure.codes}{Codes de paye à exclure pour le calcul du coût salarial (vecteur de chaînes de caractères).}

\item{exclure.sft}{Exclure le SFT du calcul}

\item{quotite.nulle}{[défaut FALSE] Si TRUE, lorsque la quotité est nulle, faire comme si elle valait 1. A n'utiliser que dans le cas de défauts de bases qui utilisent 0 pour les quotités de temps plein (attesté mais rare).}
}
\value{
Un tableau des rémunérations nettes moyennes par emploi mis en forme avec les emplois en ligne et autant de colonnes numériques que d'annees de periode, plus une colonne de libellés.
}
\description{
Elabore un tableau des rémunérations nettes moyennes du personnel par emploi et par annee (hors SFT).
}
\examples{
net.eqtp.emploi()
}
