% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{newpage}
\alias{newpage}
\title{Saut de page dans les rapports d'analyse}
\usage{
newpage()
}
\value{
Aucun
}
\description{
Saut de page dans les rapports d'analyse
}
\author{
Fabrice Nicol
}
