% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{non.null}
\alias{non.null}
\title{Sélectionne les éléments non nuls d'un vecteur}
\usage{
non.null(X)
}
\description{
Sélectionne les éléments non nuls d'un vecteur
}
\author{
Fabrice Nicol
}
