% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noria.R
\name{noria}
\alias{noria}
\title{Tableau du GVT négatif}
\usage{
noria(Bulletins = Bulletins.paie, Base = Analyse.variations.par.exercice,
  classe = "", champ = "brut", filtre = "", fichier = "", dec = ",",
  sep = ";", encoding = "UTF-8", afficher.tableau = TRUE)
}
\arguments{
\item{Bulletins}{Base mensuelle des Bulletins de paie, comportant pour l'ensemble de la période
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Année}
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Grade}                 
         \item \code{Catégorie}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{quotité}}{réel entre 0 et 1, quotité mensuelle}}}}}

\item{Base}{de paye, ou extraction de cette base, comportant pour l'ensemble de la période
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Année}
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Catégorie}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{quotité.moyenne}}{réel entre 0 et 1, quotité annuelle moyenne}
         \item{\code{Montant.net.eqtp}}{salaire net moyen annuel}
         \item{\code{Montant.brut.eqtp}}{salaire brut moyen annuel}}}}.}

\item{classe}{Valeur caractère définissant une sous-population.}

\item{champ}{"brut" ou "net" selon que le tableau est calculé sur rémunérations brutes ou nettes.}

\item{filtre}{"A", "B", "C" pour les catégories statutaires correspondantes, ou bien un vecteur de libellés de statuts possibles (exemple c("TITULAIRE", "STAGIAIRE"). Vecteur de chaînes de caractères exactes.}

\item{fichier}{Si absent, seules les bases de paye seront utilisées. 
En première année sous revue, est réputé absent le Matricule ayant une quotité non uniformément nulle sur l'année, qui n'a pas de quotité positive dans les trois premiers mois
En dernière année sous revue, est réputé absent le Matricule ayant une quotité non uniformément nulle sur l'année, qui n'a pas de quotité positive dans les trois derniers mois
Cette modélisation est adoptée par défaut si un fichier des entrées sorties n'est pas passé en paramètre fichier.
Sont exclus les vacataires et assistantes maternelles détectées.
Si présent, fichier indique le chemin complet du fichier des entrées-sorties, comportant pour l'ensemble de la période :  
       \enumerate{
         \item{ à titre obligatoire :
            \describe{
                \item{\code{Matricule}}{Matricule de l'agent, au même format que dans les autres bases}
                \item{\code{Entrée}}{Date d'entrée de l'agent en service, de la forme numérique XX/XX/20XX}                 
                \item{\code{Sortie}}{Date de sortie de l'agent du service, de la forme numérique XX/XX/20XX}}}                 
          \item{ à titre optionnel :
              \describe{
                \item{\code{Classe}}{une variable qualitative de type caractère définissant une sous-population.}}}}}

\item{dec}{Paramètre \code{dec} de la fonction \code{data.table::fread}.}

\item{sep}{Paramètre \code{sep} de la fonction \code{data.table::fread}.}

\item{encoding}{Paramètre \code{encoding} de la fonction \code{data.table::fread}.}

\item{afficher.tableau}{Si TRUE, affiche deux tableaux correspondant à la valeur de retour sur la sortie standard.}
}
\value{
Un tableau du GVT négatif au format \code{data.table}, comportant les variables numériques suivantes (les totaux de colonnes figurent seulement dans l'affichage) :
              \describe{
                \item{\code{période}}{Années de la période sous revue, puis le libellé "Total".}
                \item{\code{effet.noria}}{cumul des économies (si négatif) ou des coûts (si positif) sur salaires (nets ou bruts) dus au remplacement des sortants par les entrants, par année, puis Total.}
                \item{\code{nent}}{nombre d'entrants, par année, puis Total.}
                \item{\code{nsort}}{nombre de sortants, par année, puis Total.}
                \item{\code{variation.effectifs}}{différence du nombre d'entrants et de sortants, par année, puis Total.}
                \item{\code{taux.rotation}}{ratio égal à la moyenne du nombre d'entrants et de sortants rapportée à l'effectifs de l'année précédente. Cet effectif est estimé comme la somme des ETPT de l'année en cours, diminuée de la variation d'effectifs.}
                \item{\code{effet.variation.effectifs}}{cumul des économies (si négatif) ou des coûts (si positif) sur salaires (nets ou bruts) dus aux variations d'effectifs. Est positif si le nombre d'entrants excède le nombre de sortants, par année, puis Total.}
                \item{\code{effet.vacances}}{cumul des économies (si négatif) ou des coûts (si positif) sur salaires (nets ou bruts) dus aux vacances d'emplois. Est positif si les entrants intègrent l'organisme en moyenne plus tôt dans l'année que les sortants ne le quittent, par année, puis Total.}                                  
                \item{\code{effet.total.entrées.sorties}}{cumul des effets précédents, par année, puis Total.}}
}
\description{
Elabore un tableau de la décomposition du GVT négatif (effet de noria, de variation d'effectifs et de vacances d'emploi)
}
\examples{
noria(champ = "net", filtre = "A", classe = "SPP")
}
