% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{quotites}
\alias{quotites}
\title{Calcul des quotites}
\usage{
quotites()
}
\arguments{
\item{Bulletins.paie}{Fichier des bulletins de paye}

\item{Paie}{Base des lignes de paye}
}
\description{
Calcul des quotites
}
