% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rémunérations_eqtp}
\alias{rémunérations_eqtp}
\title{Calculer la rémunération nette et brute en EQTP}
\usage{
r攼㸹mun攼㸹rations_eqtp(DT)
}
\arguments{
\item{DT}{data.table (Bulletins.paie)}
}
\description{
Calculer la rémunération nette et brute en EQTP
}
