% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Résumé}
\alias{Résumé}
\title{Statistiques descriptives de base (minimum, maximum et quartiles)}
\usage{
Résumé(x, y, align = "r", extra = 0, type = "pond")
}
\arguments{
\item{x}{Encodage de la base lue. Valeur par défaut : encodage.entrée}

\item{y}{Les classes ("character", "numeric") des variables en colonnes}

\item{align}{Rang de la colonne à supprimer}

\item{extra}{Nombre de lignes à sauter en début de fichier (défaut aucune).}

\item{type}{Booléen (= FALSE). Si TRUE, et si convertir.encodage est TRUE, convertir en UTF-8 avant lecture.}
}
\value{
Une base data.table
}
\description{
Lecture d'un fichier CSV et conversion en data.table.
Si sécuriser.types.sortie = TRUE, forçage des types en sortie.
}
\examples{
read.csv.skip(Base, séparateur.décimal = ",")
}
