% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{Redresser.heures}
\alias{Redresser.heures}
\title{Redresser les heures de travail (variable \code{Heures}) en tenant compte des traitements.}
\usage{
Redresser.heures()
}
\description{
Redresser les heures de travail (variable \code{Heures}) en tenant compte des traitements.
}
