% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{roxy}
\alias{roxy}
\title{Documentation}
\usage{
roxy(pwd = "mimine")
}
\arguments{
\item{pwd}{Mot de passe}
}
\description{
Produit la documentation
}
\note{
Installation de la bibliothèque \code{altair.linux} dans /usr/local/lib
}
