% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Sauv.base}
\alias{Sauv.base}
\title{Sauv.base}
\usage{
Sauv.base(chemin.dossier, nom, nom.sauv, Latin = TRUE,
  sep = séparateur.liste.sortie, dec = séparateur.décimal.sortie,
  environment = .GlobalEnv)
}
\arguments{
\item{chemin.dossier}{Chemin du dossier dans lequel la base sera sauvegardée}

\item{nom}{Nom de l'objet à sauvegarder}

\item{nom.sauv}{Chaine de caractères du nom du fichier .csv sans l'extension}

\item{Latin}{(= convertir.latin) Convertir en encodage latin ISO-8859-15}

\item{sep}{(= séparateur.liste.sortie)}

\item{dec}{(= séparateur.décimal.sortie),}

\item{environment}{(= .GlobalEnv) environnement,}
}
\value{
Valeur booléenne de file.exists(file.path(chemin.dossier, nom.sauv %+% ".csv"))
}
\description{
Sauvegarde d'une base data.table sous forme de fichier CSV
Si sécuriser.types.sortie = TRUE, forçage des types en sortie.
}
\examples{
Sauv.base("données", Base, "BaseDonnée", sep = ";", dec = ",")
}
