% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{sauv.bases}
\alias{sauv.bases}
\title{Sauvegarde de plusieurs bases}
\usage{
sauv.bases(chemin.dossier, env, ...)
}
\arguments{
\item{chemin.dossier}{Chemin du dossier dans lequel la base sera sauvegardée}

\item{env}{Environnement}

\item{...}{Autres noms d'objets à sauvegarder}

\item{nom}{Nom de l'objet à sauvegarder}
}
\value{
Liste de booléens résultant de l'application de \link{Sauv.base}
}
\description{
Sauvegarde d'une ou plusieurs base de type data.table sous forme de fichier CSV
}
\examples{
envir <- environment()
# Générer Bulletins.paie et Paie dans envir
sauv.bases("données",  env = envir, c("Bulletins.paie", "Paie"))
}
\author{
Fabrice Nicol
}
