% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{sauvebase}
\alias{sauvebase}
\title{Sauvegarde une base dans le dossier des bases}
\usage{
sauvebase(x, y, z, env)
}
\arguments{
\item{x}{Objet à sauvegarder (vecteur de caractères)}

\item{y}{Nom du fichier de sauvegarde CSV.}

\item{z}{Nom du sous-dossier du dossier des bases.}

\item{env}{Environnement}
}
\description{
Sauvegarde paramétrée par environnement
}
