% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{sauvegarder.pile.bases}
\alias{sauvegarder.pile.bases}
\title{Sauvegarder les objets dont les caractéristiques ont été empilées dans \code{pile.bases}}
\usage{
sauvegarder.pile.bases()
}
\description{
Sauvegarder les objets dont les caractéristiques ont été empilées dans \code{pile.bases}
}
