% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Tableau}
\alias{Tableau}
\title{Tableau}
\usage{
Tableau(x, ...)
}
\arguments{
\item{x}{Vecteur de nom de lignes pour le tableau}

\item{...}{paramètres. Si sep.milliers est un paramètre, sa valeur est utiisée comme separateur des milliers. Par défault, le separateur blanc.}
}
\value{
Base de données data.table mise en forme de tableau par la fonction knitr::kable comportant l'ensemble des paramètres mis en colonnes centrees, avec x comme noms de lignes.
}
\description{
Présentation de tables sous forme de tableau d'une seule ligne de données.
Le tableau ne peut pas présenter plus d'une seule ligne.
}
\examples{
Tableau(c("a", "b", "c", "d"), 1, 2, 3, 4)

| a | b | c | d |
|:-:|:-:|:-:|:-:|
| 1 | 2 | 3 | 4 |

}
