% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{Tableau.vertical}
\alias{Tableau.vertical}
\title{Tableau de plusieurs lignes}
\usage{
Tableau.vertical(colnames, rownames, extra = "", ...)
}
\arguments{
\item{colnames}{Vecteur de nom de colonnes pour le tableau}

\item{rownames}{Vecteur de nom de lignes pour le tableau}

\item{extra}{(= "") Si la valeur de ce paramètre est "variation", la variation relative est calculée entre le début et la fin de la période en lignes}

\item{...}{paramètres fonctionnels uniquement. Si un paramètre n'est pas une fonction, le tableau est vide dans son ensemble}
}
\value{
Base de données data.table mise en forme de tableau par la fonction knitr::kable comportant l'ensemble des paramètres mis en colonnes centrées, avec x comme noms de lignes.
}
\description{
Présentation de tables sous forme de tableau de plusieurs lignes de données.
}
\examples{
Tableau(c("a", "b", "c", "d"), 1, 2, 3, 4)

| a | b | c | d |
|:-:|:-:|:-:|:-:|
| 1 | 2 | 3 | 4 |

}
