% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{tableau_cumuls}
\alias{tableau_cumuls}
\title{Affichage du tableau des cumuls de primes}
\usage{
tableau_cumuls(résultat)
}
\arguments{
\item{résultat}{Résultat retourné par la fonction \link{analyse}}
}
\description{
Affichage du tableau des cumuls de primes
}
