% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{tableau_NAS}
\alias{tableau_NAS}
\title{Affichage du tableau des cumuls de primes et du logement par NAS}
\usage{
tableau_NAS(résultat)
}
\arguments{
\item{résultat}{Résultat retourné par la fonction \link{test_prime}}
}
\description{
Affichage du tableau des cumuls de primes et du logement par NAS
}
\examples{
tableau_NAS(test_prime(prime_IFTS, Paie_I, verbeux = FALSE, NAS = "non"))

|Matricule |Annee | Mois |       Grade            |       Emploi            |    Montant |
|----------|------|-------------------------------|-------------------------|------------|
|010843    |2009  |   8  |  ATTACHE PRINCIPAL     |   CHEF DE DIVISION      |    785,25  |
|010843    |2009  |   9  |  ATTACHE PRINCIPAL     |   CHEF DE DIVISION      |    785,25  |
|010854    |2009  |   9  |  ADMINISTRATEUR        |   CHEF DE SERVICE       |    995,20  |

}
\seealso{
Other Tableau de primes: \code{\link{tableau_cumuls}}
}
\concept{Tableau de primes}
