% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_avn}
\alias{test_avn}
\title{Teste les logements par NAS}
\usage{
test_avn(avantage, Paie, logements = NULL)
}
\arguments{
\item{avantage}{Vecteur de caractères indiquant le type de logement (actuellement seul "NAS" est actif)}

\item{Paie}{Base de Paye principale comportant les variables \code{Statut, Grade, Emploi, Type, Code, Libellé, Montant}}

\item{logements}{La base de logements facultative importée par l'onglet Extra de l'interface graphique}
}
\value{
verbeux
}
\description{
Teste les logements par NAS
}
