% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygendoc.R
\name{toBar}
\alias{toBar}
\alias{notToBar}
\alias{theQuestion}
\title{Group of functions page title}
\usage{
toBar(x)

notToBar(x, y)

theQuestion(z)
}
\arguments{
\item{x}{a param for toBar and notToBar}

\item{y}{a param just for notToBar}

\item{z}{a param just for theQuestions}

\item{z}{a param just for theQuestion. Could put here, but easiest if all params are
described in the same place, with page documentation, so none get repeated.}
}
\value{
Hard to have one return section for all functions, might want to have a manual list here,
But if doing a manual list, that kind of avoids the purpose of the Functions section. Probably
want to describe in aggregate here and in specific with each function.
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{Functions}{
\itemize{
\item \code{toBar}: Description for the base function in the group.

   Returns no value. Not indented.

\item \code{notToBar}: Description for another function in the toBar group

\item \code{theQuestion}: Final function in toBar group. Still on same line with name
even though new paragraph. Indenting does nothing here either.
}}

\section{After function section}{

Despite its location, this actually comes after the function section.
}

\section{Another section after function section}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

