% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.données.R
\name{type.données}
\alias{type.données}
\title{Contrôle le type des données d'entrée}
\usage{
type.données(colonnes)
}
\arguments{
\item{colonnes}{Vecteur des noms de colonnes de la table globale}
}
\description{
Contrôle le type des données d'entrée
}
