% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{évolution_agrégat}
\alias{évolution_agrégat}
\title{Affichage du tableau des variations des agrégats des primes A et B sur l'ensemble de la periode}
\usage{
évolution_agrégat(résultat, verbeux)
}
\arguments{
\item{résultat}{Résultat retourné par la fonction \link{test_prime}}

\item{verbeux}{[FALSE] Le résultat n'est affiché que si  \code{verbeux} vaut  \code{TRUE}}
}
\description{
Affichage du tableau des variations des agrégats des primes A et B sur l'ensemble de la periode
}
