% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{conditionnel}
\alias{conditionnel}
\title{Insertion conditionnelle de texte dans le rapport}
\usage{
conditionnel(msg = "", path = "", DT = NULL)
}
\arguments{
\item{msg}{Partie du rapport (commençant par #')}

\item{path}{Chemin local de la base CSV sur le dossier d'exportation}

\item{DT}{base de donnée (data.rame ou data.table)}
}
\value{
Le message msg dans le rapport
}
\description{
Insertion conditionnelle de texte dans le rapport
}
