% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{FR}
\alias{FR}
\title{Ajoute le séparateur des milliers français}
\usage{
FR(x)
}
\arguments{
\item{x}{Vecteur de valeurs numériques}
}
\value{
Vecteur de chaînes de caractères modifiées
}
\description{
Ajoute le séparateur des milliers français
}
