% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliotheque.fonctions.paie.R
\name{insérer_script}
\alias{insérer_script}
\title{Insérer un script auxiliaire, indexé par une variable globale}
\usage{
insérer_script(chemin = NULL, index = c(0), variable = "année",
  gen = générer.rapport, incrémenter = FALSE, fonction = NULL,
  type = séquentiel)
}
\arguments{
\item{chemin}{Chemin du script R}

\item{index}{Vecteur numérique contenant les valeurs de la variable globale.}

\item{variable}{Vecteur de caractères contenant le nom de la variable globale dans le script auxiliaire.}

\item{gen}{Si \code{FALSE} alors se contente de sourcer le script auxiliaire selon \code{encodage.code.source}. Sinon intègre le rapport auxiliaire au format du rapport principal.}
}
\value{
Valeur de la dernière variable globale \code{variable} instanciée. Effets de bord en sortie.
}
\description{
Insérer un script auxiliaire, indexé par une variable globale
}
