% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramides.R
\name{produire_pyramides}
\alias{produire_pyramides}
\title{Produire les pyramides des âges par versant de la fonction publique}
\usage{
produire_pyramides(Filtre_bulletins, titre, versant = "")
}
\arguments{
\item{Filtre_bulletins}{Fonction permettant de filtrer les bulletins sur les lignes de data.table}

\item{titre}{Titre de la pyramide.}

\item{versant}{Versant de la fonction publique ("FPT" ou "FPH")}
}
\description{
Produire les pyramides des âges par versant de la fonction publique
}
