% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpt_rmpp.R
\name{smpt}
\alias{smpt}
\title{SMPT}
\usage{
smpt(Filtre, type = "smpt net")
}
\arguments{
\item{Filtre}{Filtre fonctionnel booléen décrivant une sélection sur les lignes (caractéristiques des personnels)}
}
\value{
Tableau d'évolution du SMPT
}
\description{
SMPT
}
